requirements.txt
mysql-connector-python
werkzeug